Reorganized the Services Namespace
