+++ src/app/Services/ContentGenerators/BaseContentService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services\Content;
+namespace App\Services\ContentGenerators;
 
 use Illuminate\Database\Eloquent\Model;
