+++ src/app/Services/ContentImporters/BibleImportService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services\ImportContent;
+namespace App\Services\ContentImporters;
 
 use App\Models\Bible;
 use App\Traits\Messageble;
