+++ src/app/Services/ContentImporters/BirthdaysImportService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services\ImportContent;
+namespace App\Services\ContentImporters;
 
 use App\Models\Birthday;
 use App\Traits\Messageble;
