+++ src/app/Services/ContentImporters/ImportContentOrchestratorService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services\ImportContent;
+namespace App\Services\ContentImporters;
 
 use Intervention\Image\Exception\InvalidArgumentException;
