+++ src/app/Services/ContentImporters/WordsImportServices.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services\ImportContent;
+namespace App\Services\ContentImporters;
 
 use App\Models\Word;
 use App\Traits\Messageble;
